Relocate IHostUiHandler to Common.HostUiBridge
